Update main.css
I used fractional units to define how large the grid is. Therefore, for the header it is only 1 fractional unit, as there is only one section used. The main container is split up into 4 fractional units, even though there is 3 sections, as the main container should take up twice the space that the navigation and aside takes up. Lastly, the footer container is only one fractional unit, as there is only one section. A column gap was given to the main container, as there was not enough space between the sections.